Fixed last value not checked for min_int or max_int

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: lmuzio <lmuzio@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/19 18:54:27 by lmuzio            #+#    #+#             */
-/*   Updated: 2022/10/13 16:56:22 by lmuzio           ###   ########.fr       */
+/*   Updated: 2022/10/13 17:54:05 by lmuzio           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@
 # define RRB 14
 # define RRR 15
 
-int				check_duplicates(t_list *a, int value);
+int				check_duplicates(t_list *a, long value);
 int				check_digits(const char *str);
 int				find_max_bit(t_list *stack);
 t_list			*parse_arguments(int argc, char *argv[], t_list *a);
